backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.35
python-multipart==0.0.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.0
pydantic==2.9.2
pillow==10.4.0
aiofiles==24.1.0
itsdangerous==2.2.0
typing-extensions==4.12.2
numpy>=2.0.0
google-generativeai>=0.8.3

